refactor: activate마커 핸들링 관련 코드 리팩토링
